wiener index is a distance measure
